// src/common/types.ts
// NOTE! The values on the right need to be an exact match to the command names a user types.
export enum Commands {
  CreateContext = 'create-context',
  CurrentContext = 'current-context',
  DeleteContext = 'delete-context',
  UseContext = 'use-context',
  Config = 'config',
  Reviewer = 'reviewer',
  Main = 'main',
  Open = 'open',
}